requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django
djangorestframework
idna==3.2
pytz==2021.1
requests
sqlparse==0.4.1
urllib3==1.26.6
django-mysql==3.8.1
mysqlclient
